Move Torznab API to /api/torznab/api path for Prowlarr compatibility
