+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "trackarr",
-  "version": "0.5.6",
+  "version": "0.5.7",
   "type": "module",
   "private": true,
   "scripts": {
